+++ R/data_part.R
@@ -11,11 +11,12 @@
 ################################################################################
 ################################################################################
 ################################################################################ 
+# function 1
 data_part <- function(data, partition=2L, probs, setseed=123, ...)
 {
   set.seed(setseed)
 if (partition<=1L) stop("data partition should be greater that one","\n")
-if (partition>=20L) stop("data partition should be less that 20","\n")
+if (partition>=21L) stop("data partition should be less or equal that 20","\n")
 if (partition==2L)
 {
   cat("data partition into two sets", "\n")
@@ -54,18 +55,19 @@ if (partition>=4L)
 ################################################################################
 ################################################################################
 ################################################################################
-# This function greates a list with different partions of the same data 
-# no more tahn 3 partitions
+# This function creates a list with different partions of the same data 
+# no more than 3 partitions
 ################################################################################
 ################################################################################
 ################################################################################
 ################################################################################
-data_part_list <- function(data, partition=2, probs, setseed=123, ...)
+# Function 2
+data_part_list <- function(data, partition=2L, probs, setseed=123, ...)
 {
   set.seed(setseed)
   if (partition<=1L) stop("data partition should be greater that one","\n")
-  if (partition>=20L) stop("data partition should be less that 20","\n")
-  if (partition!=2&&partition!=3) stop("partition should be 2 or 3")
+  if (partition >= 21L) stop("data partition should be less or equal that 20","\n")
+ # if (partition!=2&&partition!=3) stop("partition should be 2 or 3")
   if(missing(probs))
   {
     probs <- if (partition==2)  c(0.6,0.4) 
@@ -74,13 +76,14 @@ data_part_list <- function(data, partition=2, probs, setseed=123, ...)
   
   if (length(probs)>4||length(probs)<=0) stop("the length of probs should be  2 o 3")
   if (sum(probs)!=1) stop("probs should add up to 1")
-  if (partition==2)
+  if 
+  (partition==2)
   {
      rand <- sample(2, dim(data)[1], replace=TRUE, prob=probs)
     train <- subset(data, rand==1)
      test <- subset(data, rand==2)
       out <- list(train=train , test=test)
-    return(out)
+      invisible(return(out))
   }
   if (partition==3)
   {
@@ -89,89 +92,141 @@ data_part_list <- function(data, partition=2, probs, setseed=123, ...)
     valid <- subset(data, rand==2)
      test <- subset(data, rand==3)
       out <- list(train=train, valid=valid, test=test)
-    return(out)              
+      invisible(return(out))              
   }  
-}
-################################################################################
-################################################################################
-################################################################################
-################################################################################
-data_part_index <- function(data, K=2, bootstrap=FALSE)
-{
-  dD <- dim(data)
-  N <- dim(data)[1]
-  if (bootstrap)  
+  if (partition>=4L)
   {
-    mm <-    lapply(
-               as.data.frame(
-               t(
-                 sapply(
-                   sample(rep_len(1:N, length.out=N), replace=TRUE),
-                   "!=",1:K)
-                 )
-                            ), 
-                        which)    
-  } else 
-  {
-    mm <-    lapply(
-              as.data.frame(
-                t(
-                  sapply(
-                    sample(rep_len(1:K, length.out= N), replace=FALSE)
-                        ,"!=", 1:K )
-                  )
-                          ), 
-              which ) 
+    cat( paste0(partition,"-fold data partition"), "\n")
+      out <- list()
+    probs <- rep(1/partition, partition)
+    if (abs(sum(probs)-1)>1.0e-10) stop("probs should add up to 1")    
+     rand <- sample(partition, dim(data)[1], replace=TRUE, prob=probs)
+     rand <- factor(rand)
+for (i in 1:partition)
+    {
+      out[[i]] <- subset(data, rand==i)  
+    }
+    invisible(return(out))   
   } 
-  mm
 }
+################################################################################
+################################################################################
+################################################################################
+################################################################################
+# I am not sure about that at the moment I may have to go out
+# data_part_index <- function(data, K=10, bootstrap=FALSE)
+# {
+#   dD <- dim(data)
+#   N <- dim(data)[1]
+#   if (bootstrap)  
+#   {
+#     mm <-    lapply(
+#                as.data.frame(
+#                t(
+#                  sapply(
+#                    sample(rep_len(1:N, length.out=N), replace=TRUE),
+#                    "!=",1:K)
+#                  )
+#                             ), 
+#                         which)    
+#   } else 
+#   {
+#     mm <-    lapply(
+#               as.data.frame(
+#                 t(
+#                   sapply(
+#                     sample(rep_len(1:K, length.out= N), replace=FALSE)
+#                         ,"!=", 1:K )
+#                   )
+#                           ), 
+#               which ) 
+#   } 
+#   mm
+# }
 ################################################################################
 ################################################################################
 ################################################################################
 ################################################################################
 # I was not sure if this function is working so it is not documented in the package 
-data_Kfold <- function(data, K=6, setseed=123 )
+data_Kfold <- function(data, K = 10, setseed=123)
 {
   set.seed(setseed)
   n <- dim(data)[1]
   # folds for cross-validation 
-  CVfolds <-  lapply(
-    as.data.frame(
+  CVfolds <-
       t(
         sapply(
           sample(rep_len(1:K, length.out=n),replace=FALSE)
           ,"!=", 1:K)
-      )
-    )
-    , which )   
-  CVfolds
+              )*1 #if you want it numeric
+  return(CVfolds)
 }
 ################################################################################
 ################################################################################
 ################################################################################
 ################################################################################
 # I was not sure if this function is working so it is not documented in the package 
+# Mikis 22-4-25 It look OK to me 
 # get a data frame and creates a K bootstrap indices as a matrix or a list
-data_boot <- function(data, K=10, setseed=123 ,as.matrix=TRUE)
+data_boot_index <- function(data, K=10, setseed=123 , as.matrix=TRUE)
 {
-  set.seed(setseed)
-  nfolds <- K
-  n <- dim(data)[1]
-  BOOTfolds<- lapply( 
+set.seed(setseed)
+     nfolds <- K
+          n <- dim(data)[1]
+  BOOTfolds <- lapply( 
      as.data.frame(
          matrix(
-        sample(1:n, nfolds*n, replace=TRUE)
+        sample(1:n, nfolds*n,replace=TRUE)
         , n)
     ),
     sort) 
-  if (as.matrix) return(matrix(unlist(BOOTfolds), nrow=n))
-  else return(BOOTfolds)
+  BOOTnotfolds <- list()
+  for (i in 1:K)
+    BOOTnotfolds[[i]] <-  setdiff(1:n,BOOTfolds[[i]]) 
+ return(list(IB=BOOTfolds,OOB=BOOTnotfolds))  
+  # if (as.matrix) return(matrix(unlist(BOOTfolds), nrow=n))
+  # else return(BOOTfolds)
 }
 ################################################################################
 ################################################################################
 ################################################################################
 ################################################################################
-
+data_boot_weights <- function(data, K=10, setseed=123 , as.matrix=TRUE)
+{
+  set.seed(setseed)
+  nfolds <- K
+  n <- dim(data)[1]
+  BOOTfolds <- lapply( 
+    as.data.frame(
+      matrix(
+        sample(1:n, nfolds*n,replace=TRUE)
+        , n)
+    ),
+    sort) 
+  BOOTnotfolds <- list()
+  for (i in 1:K)
+    BOOTnotfolds[[i]] <-  setdiff(1:n,BOOTfolds[[i]]) 
+  return(list(IB=BOOTfolds,OOB=BOOTnotfolds))  
+  # if (as.matrix) return(matrix(unlist(BOOTfolds), nrow=n))
+  # else return(BOOTfolds)
+}
+# get_kfolds <- function(data, K=6, setseed=123 )
+# {
+#   set.seed(setseed)
+#   nfolds <- K
+#   n <- dim(data)[1]
+#   # folds for cross-validation 
+#   CVfolds <-  lapply(
+#     as.data.frame(
+#       t(
+#         sapply(
+#           sample(rep_len(1:nfolds,length=n),replace=FALSE)
+#           ,"!=", 1:nfolds)
+#       )
+#     )
+#     , which )   
+#   CVfolds
+# }
 
 
 
@@ -274,8 +329,12 @@ data_boot <- function(data, K=10, setseed=123 ,as.matrix=TRUE)
 #   }  
 # }
 
+#CVfolds<- lapply(as.data.frame(t(sapply(sample(rep_len(1:nfolds,length=n),replace=FALSE)
+#                                        ,"!=", 1:nfolds))), which)  
+#BOOTfolds<- lapply(as.data.frame(matrix(sample(1:n, nfolds*n, replace=TRUE), n)),sort)  
 
-
+#BOOTfolds<- lapply(as.data.frame(matrix(sample(1:10, 10*10, replace=TRUE), 19)),sort) 
+#matrix(unlist(BOOTfolds),nrow=10)
 ########
 
 #data_rm(rent99, c("rentsqm", "district"))